fix all tests for jwt
