+++ pallets/funding/Cargo.toml
@@ -22,7 +22,6 @@ tokio = { version = "1", features = ["full"], optional = true }
 futures = { version = "0.3.29", optional = true }
 log.workspace = true
 itertools.workspace = true
-jwt-compact = { workspace = true, features = ["ed25519-dalek"] }
 
 # Substrate dependencies
 frame-benchmarking = {workspace = true, optional = true}
@@ -54,14 +53,14 @@ polimec-receiver.workspace = true
 # xcm-executor.workspace = { workspace = true, optional = true }
 
 [dev-dependencies]
+assert_matches2.workspace = true
 pallet-insecure-randomness-collective-flip.workspace = true
 pallet-timestamp.workspace = true
 pallet-assets.workspace = true
 pallet-linear-release.workspace = true
-assert_matches2.workspace = true
-xcm-executor.workspace = true
+polimec-common-test-utils.workspace = true
 macros.workspace = true
-reqwest = { version = "0.11.4", features = ["blocking"] }
+xcm-executor.workspace = true
 
 [features]
 default = [ "std" ]
@@ -71,7 +70,6 @@ std = [
 	"frame-system/std",
 	"futures",
 	"itertools/use_std",
-	"jwt-compact/std",
 	"log/std",
 	"pallet-assets/std",
 	"pallet-balances/std",
@@ -82,6 +80,7 @@ std = [
 	"parachains-common/std",
 	"parity-scale-codec/std",
 	"polimec-common/std",
+	"polimec-common-test-utils/std",
 	"polimec-receiver/std",
 	"polimec-xcm-executor/std",
 	"polkadot-parachain/std",
