+++ polimec-common/common/src/lib.rs
@@ -18,8 +18,8 @@
 
 use frame_support::{pallet_prelude::*, traits::tokens::fungible, RuntimeDebug};
 use itertools::Itertools;
-
 use sp_std::prelude::*;
+pub mod credentials;
 
 /// A release schedule over a fungible. This allows a particular fungible to have release limits
 /// applied to it.
